requirements.txt
flask
google-generativeai
manim
elevenlabs
python-dotenv
